// client/src/generated/PlayerStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ClassId } from "./ClassId";

export type PlayerStats = { name: string, class: ClassId, kills: number, damage: number, healing: number, xp: number, coins: number, alive: boolean, };