// client/src/generated/RunStartInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Modifiers } from "./Modifiers";
import type { PlayerInfo } from "./PlayerInfo";

export type RunStartInfo = { width: number, height: number, 
/**
 * One byte per tile, row-major. See `Tile` in dungeon/mod.rs.
 */
tiles: Uint8Array, tile_size: number, 
/**
 * Entity id of the receiving player.
 */
you: number, players: Array<PlayerInfo>, spawn_x: number, spawn_y: number, 
/**
 * Progression multipliers of the receiving player (prediction must use them).
 */
mods: Modifiers, };